travis: Ignore dependencies with travis
When building on travis with clang, libboost1.55-dev is used.

We have to set this for all compilers as libboost-dev is an incompatble
package.

As such, when building on travis, use the -d option to dpkg-buildpackage
to ignore dependencies as these get determined from the distro
rather than the debian/rules file.