MDEV-14340: mariadb.pc install should be arch dependent
As such use CMAKE_INSTALL_LIBDIR to set the path.

Despite being a move in location, this is a stable move as
this file is for the use by pkg-config. Pkg-config will pick the
file up in the new location transparently. It also will
correct cross compulation problems that against mariadb.so
because the location is now correct for the achitecture compiled.

CMAKE_INSTALL_LIBDIR was added to CMake in version 2.8.5
(ref:
https://github.com/Kitware/CMake/commit/a262fe0917364cb003f7319b3d17c46280d75927#diff-ddd40a000ca37e13b4cae5b9bd023c53R21)

Debian Wheezy has 2.8.9 (https://packages.debian.org/wheezy/cmake)
and the Centos 6 buildbot seem to be running at 2.8.12.2.